Rename module 'pint' to '_pint' in attempt to solve build failing on ubuntu, maybe because of weird namespace thing?
