+++ pdtable/units/converter.py
@@ -1,7 +1,7 @@
 """Default unit converter"""
 
 try:
-    from .pint import PintUnitConverter as DefaultUnitConverter
+    from ._pint import PintUnitConverter as DefaultUnitConverter
 except ImportError as err:
     raise ImportError(
         "Unable to import 'pint'. Please install 'pint' to use the default unit converter."
